+++ src/kitaru/__init__.py
@@ -61,6 +61,7 @@ _default_analytics_source = os.environ.get(
 )
 set_source(_default_analytics_source)
 
+from kitaru._client._models import AuthAPIKey, AuthAPIKeyWithValue, AuthServiceAccount
 from kitaru._interface_deployments import Deployment
 from kitaru.artifacts import load, save
 from kitaru.checkpoint import checkpoint
@@ -107,6 +108,9 @@ from kitaru.secrets import (
 from kitaru.wait import wait
 
 __all__ = [
+    "AuthAPIKey",
+    "AuthAPIKeyWithValue",
+    "AuthServiceAccount",
     "Deployment",
     "FailureOrigin",
     "FlowHandle",